requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.1
Faker==25.2.0
google-auth==2.29.0
gunicorn==22.0.0
idna==3.7
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclean==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
